Respect requested device for a task.
On the continuous control side:
Replay memory was always on the CPU.
If an agent was on the GPU, replay buffer would still be on CPU.
This caused exceedingly strange RuntimeErrors from pytorch.

On the classification side:
We load data from our dataloaders, but nowhere do we specify which device.
As above, this caused strange RuntimeErrors.

Our RNN code was not correctly moving cell state to the same device as its input.