Add  `device` sanity checks in replay memory.
Very frustrating bugs occur when replay memory and agents are on different devices.
These errors do not occur at the inital call site.
That is, the error does not occur when attempting to cpy from one device to another.

When copying the buffer from the CPU to the GPU, grad information is mangled.

To avoid this bug entirely, require that logs match on devices.